Move Game components into their own directories.
