+++ src/components/pages/GameContainer/Game/Finale/index.ts
@@ -1,0 +1,1 @@
+export { default as Finale } from './Finale';
