+++ src/components/pages/GameContainer/Game/Guessing/index.ts
@@ -1,0 +1,1 @@
+export { default as Guessing } from './Guessing';
