+++ src/components/pages/GameContainer/Game/Lobby/index.ts
@@ -1,0 +1,1 @@
+export { default as Lobby } from './Lobby';
