+++ src/components/pages/GameContainer/Game/Postgame/index.ts
@@ -1,0 +1,1 @@
+export { default as Postgame } from './Postgame';
