+++ src/components/pages/GameContainer/Game/Pregame/index.ts
@@ -1,0 +1,1 @@
+export { default as Pregame } from './Pregame';
