+++ src/components/pages/GameContainer/Game/Writing/index.ts
@@ -1,0 +1,1 @@
+export { default as Writing } from './Writing';
