+++ src/components/pages/GameContainer/Game/index.ts
@@ -1,8 +1,8 @@
-import PlayerList from '../../../common/PlayerList/PlayerList';
-import Guessing from './Guessing';
-import Lobby from './Lobby';
-import Postgame from './Postgame';
-import Pregame from './Pregame';
-import Writing from './Writing';
+import { Finale } from './Finale';
+import { Guessing } from './Guessing';
+import { Lobby } from './Lobby';
+import { Postgame } from './Postgame';
+import { Pregame } from './Pregame';
+import { Writing } from './Writing';
 
-export { Lobby, Pregame, PlayerList, Writing, Guessing, Postgame };
+export { Lobby, Pregame, Writing, Guessing, Postgame, Finale };
