+++ src/components/pages/GameContainer/GameContainer.tsx
@@ -23,8 +23,7 @@ import { MAX_SECONDS, REACT_APP_API_URL } from '../../../utils/constants';
 import { errorCodeChecker, randomUsername } from '../../../utils/helpers';
 import { Header } from '../../common/Header';
 import { Modal } from '../../common/Modal';
-import { Guessing, Lobby, Postgame, Pregame, Writing } from './Game';
-import Finale from './Game/Finale';
+import { Finale, Guessing, Lobby, Postgame, Pregame, Writing } from './Game';
 
 // Create a socket connection to API
 const socket = io.connect(REACT_APP_API_URL as string);
